refactor: navigation-drawer fix paths
