+++ src/components/nav-drawer/styles/indigo/nav-drawer-header-item.indigo.scss
@@ -4,5 +4,4 @@
 
 :host {
     color: color(gray, 50);
-    background: red;
 }
